Move cordova keystore related things to separate file
