+++ src/platform/cordova/key-store.ts
@@ -7,40 +7,39 @@ import { Account } from "../../context/accounts"
 export default async function createKeyStore(): Promise<KeyStoreAPI> {
   return {
     async getKeyIDs() {
-      const event = await sendCommand(commands.getKeyIDsCommand)
+      const event = await sendCommand(commands.keyStore.getKeyIDsCommand)
       return event.data.result
     },
     async getPublicKeyData(keyID: string) {
       const data = { keyID }
-      const event = await sendCommand(commands.getPublicKeyDataCommand, data)
+      const event = await sendCommand(commands.keyStore.getPublicKeyDataCommand, data)
       return event.data.result
     },
     async getPrivateKeyData(keyID: string, password: string) {
       const data = { keyID, password }
-      const event = await sendCommand(commands.getPrivateKeyDataCommand, data)
+      const event = await sendCommand(commands.keyStore.getPrivateKeyDataCommand, data)
       return event.data.result
     },
     async saveKey(keyID: string, password: string, privateData: PrivateKeyData, publicData?: PublicKeyData) {
       const data = { keyID, password, privateData, publicData }
-      const event = await sendCommand(commands.saveKeyCommand, data)
-      return event.data
+      await sendCommand(commands.keyStore.saveKeyCommand, data)
     },
     async savePublicKeyData(keyID: string, publicData: PublicKeyData) {
       const data = { keyID, publicData }
-      const event = await sendCommand(commands.savePublicKeyDataCommand, data)
-      return event.data
+      await sendCommand(commands.keyStore.savePublicKeyDataCommand, data)
     },
     async signTransaction(transaction: Transaction, walletAccount: Account, password: string) {
       const transactionEnvelope = transaction.toEnvelope().toXDR("base64")
-      const data = { transactionEnvelope, walletAccount, password }
-      const event = await sendCommand(commands.signTransactionCommand, data)
+      const stringifiedAccount = JSON.stringify(walletAccount)
+      const data = { transactionEnvelope, stringifiedAccount, password }
 
-      return new Transaction(event.data.result)
+      const event = await sendCommand(commands.keyStore.signTransactionCommand, data)
+      const signedTransactionEnvelope = event.data.result
+      return new Transaction(signedTransactionEnvelope)
     },
     async removeKey(keyID: string) {
       const data = { keyID }
-      const event = await sendCommand(commands.removeKeyCommand, data)
-      return event.data
+      await sendCommand(commands.keyStore.removeKeyCommand, data)
     }
   }
 }
